# docs/swagger.yaml
definitions:
  handler.IncomingTodo:
    properties:
      title:
        type: string
    type: object
host: 127.0.0.1:8000
info:
  contact:
    email: syahidatulamira06@gmail.com
    name: Amira
  description: This is a swagger for Todo
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Todo
  version: "1.0"
paths:
  /todo/:
    get:
      description: Get all todo items by user id from login
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all todo items by user id
      tags:
      - Todo
    post:
      consumes:
      - application/json
      description: Add a new todo item
      parameters:
      - description: Todo object
        in: body
        name: IncomingTodo
        required: true
        schema:
          $ref: '#/definitions/handler.IncomingTodo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create new todo item
      tags:
      - Todo
  /todo/{id}:
    delete:
      description: Delete a todo item by ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete todo item by ID
      tags:
      - Todo
  /todo/complete/{id}:
    patch:
      description: Update a todo item status to complete by ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a todo item status to complete
      tags:
      - Todo
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"